IVYPORTAL-14388 Welcome Widget: Allow html code in the text, enhance …

+++ AxonIvyPortal/portal/src/ch/ivy/addon/portal/generic/bean/DashboardDetailModificationBean.java
@@ -424,7 +424,7 @@ public class DashboardDetailModificationBean extends DashboardBean implements Se
     if (StringUtils.isNotBlank(welcomeWidget.getImageLocation())) {
       ContentObjectValue tempImageFile = getWelcomeWidgetImage(true, welcomeWidget);
       ContentObjectValue imageFile = getWelcomeWidgetImage(false, welcomeWidget);
-      if (imageFile != null && tempImageFile != null) {
+      if (imageFile != null && tempImageFile != null && tempImageFile.parent().exists()) {
         imageFile.write().bytes(tempImageFile.read().bytes());
         tempImageFile.delete();
       }
